+++ basic-feature/src/main/java/com/hivian_compose_mvvm/basic_feature/data/di/DatabaseModule.kt
@@ -1,7 +1,7 @@
 package com.hivian_compose_mvvm.basic_feature.data.di
 
-import com.hivian.compose_mvvm.core.datasource.local.dao.IRandomUsersDao
-import com.hivian_compose_mvvm.basic_feature.data.sources.local.RandomUsersDatabaseService
+import com.hivian.compose_mvvm.core.datasources.local.dao.IRandomUsersDao
+import com.hivian_compose_mvvm.basic_feature.data.sources.RandomUsersDatabaseService
 import com.hivian_compose_mvvm.basic_feature.domain.services.IRandomUsersDatabaseService
 import dagger.Module
 import dagger.Provides
