+++ basic-feature/src/main/java/com/hivian_compose_mvvm/basic_feature/data/di/NetworkingModule.kt
@@ -1,7 +1,7 @@
 package com.hivian_compose_mvvm.basic_feature.data.di
 
-import com.hivian.compose_mvvm.core.datasource.remote.IApiService
-import com.hivian_compose_mvvm.basic_feature.data.sources.remote.RandomUsersHttpService
+import com.hivian.compose_mvvm.core.datasources.remote.IApiService
+import com.hivian_compose_mvvm.basic_feature.data.sources.RandomUsersHttpService
 import com.hivian_compose_mvvm.basic_feature.domain.services.IRandomUsersHttpService
 import dagger.Module
 import dagger.Provides
