+++ basic-feature/src/main/java/com/hivian_compose_mvvm/basic_feature/data/sources/RandomUsersDatabaseService.kt
@@ -1,7 +1,7 @@
-package com.hivian_compose_mvvm.basic_feature.data.sources.local
+package com.hivian_compose_mvvm.basic_feature.data.sources
 
-import com.hivian.compose_mvvm.core.datasource.local.dao.IRandomUsersDao
-import com.hivian.compose_mvvm.core.datasource.models.RandomUserDTO
+import com.hivian.compose_mvvm.core.datasources.local.dao.IRandomUsersDao
+import com.hivian.compose_mvvm.core.datasources.models.RandomUserDTO
 import com.hivian_compose_mvvm.basic_feature.domain.services.IRandomUsersDatabaseService
 import javax.inject.Inject
