+++ basic-feature/src/main/java/com/hivian_compose_mvvm/basic_feature/domain/services/IRandomUsersDatabaseService.kt
@@ -1,6 +1,6 @@
 package com.hivian_compose_mvvm.basic_feature.domain.services
 
-import com.hivian.compose_mvvm.core.datasource.models.RandomUserDTO
+import com.hivian.compose_mvvm.core.datasources.models.RandomUserDTO
 
 interface IRandomUsersDatabaseService {
