+++ basic-feature/src/main/java/com/hivian_compose_mvvm/basic_feature/domain/services/IRandomUsersHttpService.kt
@@ -1,8 +1,8 @@
 package com.hivian_compose_mvvm.basic_feature.domain.services
 
-import com.hivian.compose_mvvm.core.datasource.HttpResult
-import com.hivian.compose_mvvm.core.datasource.models.Results
-import com.hivian.compose_mvvm.core.datasource.remote.IHttpService
+import com.hivian.compose_mvvm.core.datasources.remote.HttpResult
+import com.hivian.compose_mvvm.core.datasources.models.Results
+import com.hivian.compose_mvvm.core.datasources.remote.IHttpService
 
 interface IRandomUsersHttpService: IHttpService {
