+++ basic-feature/src/main/java/com/hivian_compose_mvvm/basic_feature/domain/usecases/RandomUsersUseCase.kt
@@ -1,6 +1,6 @@
 package com.hivian_compose_mvvm.basic_feature.domain.usecases
 
-import com.hivian.compose_mvvm.core.datasource.ServiceResult
+import com.hivian.compose_mvvm.core.datasources.ServiceResult
 import com.hivian_compose_mvvm.basic_feature.domain.models.RandomUser
 import com.hivian_compose_mvvm.basic_feature.domain.repository.IRandomUsersRepository
 import javax.inject.Inject
