+++ basic-feature/src/test/java/com/hivian_compose_mvvm/basic_feature/presentation/HomeViewModelTest.kt
@@ -3,12 +3,12 @@ package com.hivian_compose_mvvm.basic_feature.presentation
 import com.hivian.compose_mvvm.core.InstantExecutorExtension
 import com.hivian.compose_mvvm.core.MainCoroutineExtension
 import com.hivian.compose_mvvm.core.base.ViewModelVisualState
-import com.hivian.compose_mvvm.core.datasource.ErrorType
-import com.hivian.compose_mvvm.core.datasource.ServiceResult
-import com.hivian.compose_mvvm.core.datasource.models.Location
-import com.hivian.compose_mvvm.core.datasource.models.Name
-import com.hivian.compose_mvvm.core.datasource.models.Picture
-import com.hivian.compose_mvvm.core.datasource.models.RandomUserDTO
+import com.hivian.compose_mvvm.core.datasources.remote.ErrorType
+import com.hivian.compose_mvvm.core.datasources.ServiceResult
+import com.hivian.compose_mvvm.core.datasources.models.Location
+import com.hivian.compose_mvvm.core.datasources.models.Name
+import com.hivian.compose_mvvm.core.datasources.models.Picture
+import com.hivian.compose_mvvm.core.datasources.models.RandomUserDTO
 import com.hivian.compose_mvvm.core.services.ILocalizationService
 import com.hivian_compose_mvvm.basic_feature.data.mappers.ImageSize
 import com.hivian_compose_mvvm.basic_feature.data.mappers.mapToRandomUsers
