+++ core/src/main/java/com/hivian/compose_mvvm/core/datasources/local/AppDatabase.kt
@@ -1,15 +1,15 @@
-package com.hivian.compose_mvvm.core.datasource.local
+package com.hivian.compose_mvvm.core.datasources.local
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.hivian.compose_mvvm.core.datasource.local.converters.LocationConverter
-import com.hivian.compose_mvvm.core.datasource.local.converters.NameConverter
-import com.hivian.compose_mvvm.core.datasource.local.converters.PictureConverter
-import com.hivian.compose_mvvm.core.datasource.local.dao.IRandomUsersDao
-import com.hivian.compose_mvvm.core.datasource.models.RandomUserDTO
+import com.hivian.compose_mvvm.core.datasources.local.converters.LocationConverter
+import com.hivian.compose_mvvm.core.datasources.local.converters.NameConverter
+import com.hivian.compose_mvvm.core.datasources.local.converters.PictureConverter
+import com.hivian.compose_mvvm.core.datasources.local.dao.IRandomUsersDao
+import com.hivian.compose_mvvm.core.datasources.models.RandomUserDTO
 
 @Database(entities = [RandomUserDTO::class], version = AppDatabase.DB_VERSION, exportSchema = false)
 @TypeConverters(NameConverter::class, LocationConverter::class, PictureConverter::class)
