+++ core/src/main/java/com/hivian/compose_mvvm/core/datasources/local/converters/LocationConverter.kt
@@ -1,19 +1,19 @@
-package com.hivian.compose_mvvm.core.datasource.local.converters
+package com.hivian.compose_mvvm.core.datasources.local.converters
 
 import androidx.room.TypeConverter
-import com.hivian.compose_mvvm.core.datasource.models.Picture
+import com.hivian.compose_mvvm.core.datasources.models.Location
 import com.hivian.compose_mvvm.core.extensions.fromJson
 import com.hivian.compose_mvvm.core.extensions.toJson
 
-class PictureConverter {
+class LocationConverter {
 
     @TypeConverter
-    fun pictureToJson(value: Picture): String {
+    fun locationToJson(value: Location): String {
         return value.toJson()
     }
 
     @TypeConverter
-    fun jsonToPicture(value: String): Picture {
+    fun jsonToLocation(value: String): Location {
         return value.fromJson()
     }
 }
