+++ core/src/main/java/com/hivian/compose_mvvm/core/datasources/local/converters/NameConverter.kt
@@ -1,7 +1,7 @@
-package com.hivian.compose_mvvm.core.datasource.local.converters
+package com.hivian.compose_mvvm.core.datasources.local.converters
 
 import androidx.room.TypeConverter
-import com.hivian.compose_mvvm.core.datasource.models.Name
+import com.hivian.compose_mvvm.core.datasources.models.Name
 import com.hivian.compose_mvvm.core.extensions.fromJson
 import com.hivian.compose_mvvm.core.extensions.toJson
