+++ core/src/main/java/com/hivian/compose_mvvm/core/datasources/local/converters/PictureConverter.kt
@@ -1,19 +1,19 @@
-package com.hivian.compose_mvvm.core.datasource.local.converters
+package com.hivian.compose_mvvm.core.datasources.local.converters
 
 import androidx.room.TypeConverter
-import com.hivian.compose_mvvm.core.datasource.models.Location
+import com.hivian.compose_mvvm.core.datasources.models.Picture
 import com.hivian.compose_mvvm.core.extensions.fromJson
 import com.hivian.compose_mvvm.core.extensions.toJson
 
-class LocationConverter {
+class PictureConverter {
 
     @TypeConverter
-    fun locationToJson(value: Location): String {
+    fun pictureToJson(value: Picture): String {
         return value.toJson()
     }
 
     @TypeConverter
-    fun jsonToLocation(value: String): Location {
+    fun jsonToPicture(value: String): Picture {
         return value.fromJson()
     }
 }
